requirements.txt
torch==2.10.0
transformers==4.41.2
# transformers==4.38.2
xformers>=0.0.21
timm>=0.9.16
accelerate==0.33.0
datasets
sentencepiece
attrdict
einops
peft==0.11.1
bitsandbytes
pymupdf
pillow
gradio
rapidfuzz
tqdm
requests
beautifulsoup4
lxml
sentence-transformers
faiss-cpu
# faiss-gpu